+++ Tests/SentryTests/Integrations/SentrySwiftIntegrationInstallerTests.swift
@@ -19,6 +19,7 @@ final class SentrySwiftIntegrationInstallerTests: XCTestCase {
         options.swiftAsyncStacktraces = true
         options.enableAutoSessionTracking = false
         options.enableSwizzling = false
+        options.experimental.enableMetrics = false
         
         let testHub = TestHub(client: nil, andScope: nil)
         SentrySDKInternal.setCurrentHub(testHub)
@@ -40,6 +41,7 @@ final class SentrySwiftIntegrationInstallerTests: XCTestCase {
         options.swiftAsyncStacktraces = false
         options.enableAutoSessionTracking = false
         options.enableSwizzling = false
+        options.experimental.enableMetrics = false
         
         let testHub = TestHub(client: nil, andScope: nil)
         SentrySDKInternal.setCurrentHub(testHub)